pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "binarybyte"
version = "0.1.0"
description = "Infrastructure layer for AI coding agents — evaluate, sync, and deploy agent configs across tools."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "BinaryByte Contributors" },
]
keywords = ["ai", "agents", "cli", "evaluation", "deployment", "infrastructure"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Build Tools",
]
dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.4.0",
]

[project.scripts]
binarybyte = "binarybyte.cli.main:app"

[project.urls]
Homepage = "https://github.com/BinaryByte-project/BinaryByte"
Repository = "https://github.com/BinaryByte-project/BinaryByte"
Issues = "https://github.com/BinaryByte-project/BinaryByte/issues"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
